RavenDB-18134 Sharded bulk insert base implementation and refactoring of batches infrastructure handling
