+++ src/Raven.Client/Documents/Operations/OperationMultipleExceptionsResult.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Sparrow.Json.Parsing;
+
+namespace Raven.Client.Documents.Operations;
+
+public class OperationMultipleExceptionsResult : IOperationResult
+{
+    public OperationMultipleExceptionsResult()
+    {
+        
+    }
+
+    public OperationMultipleExceptionsResult(string message)
+    {
+        Message = message;
+    }
+
+    public string Message { get; }
+
+    public List<OperationExceptionResult> Exceptions { get; } = new();
+
+    public DynamicJsonValue ToJson()
+    {
+        return new DynamicJsonValue()
+        {
+            [nameof(Message)] = Message,
+            [nameof(Exceptions)] = new DynamicJsonArray(Exceptions.Select(x => x.ToJson()))
+        };
+    }
+
+    public bool ShouldPersist { get; }
+}
