+++ src/Raven.Server/Documents/Handlers/Batches/BatchRequestParser.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.Buffers;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.WebUtilities;
-using Microsoft.Net.Http.Headers;
 using Newtonsoft.Json;
 using Raven.Client.Documents.Attachments;
 using Raven.Client.Documents.Commands.Batches;
@@ -16,26 +13,23 @@ using Raven.Client.Documents.Operations.Attachments;
 using Raven.Client.Documents.Operations.Counters;
 using Raven.Client.Documents.Operations.TimeSeries;
 using Raven.Client.Documents.Session;
-using Raven.Client.Documents.Smuggler;
-using Raven.Client.Util;
+using Raven.Server.Documents.Handlers.Batches.Commands;
 using Raven.Server.Documents.Patch;
-using Raven.Server.Documents.Sharding.Handlers;
+using Raven.Server.Documents.Sharding.Handlers.Batches;
 using Raven.Server.Documents.TransactionCommands;
 using Raven.Server.Exceptions;
-using Raven.Server.ServerWide;
-using Raven.Server.ServerWide.Context;
-using Raven.Server.Smuggler;
-using Raven.Server.Web;
 using Sparrow;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Utils;
 using PatchRequest = Raven.Server.Documents.Patch.PatchRequest;
 
-namespace Raven.Server.Documents.Handlers
+namespace Raven.Server.Documents.Handlers.Batches
 {
-    public static class BatchRequestParser
+    public class BatchRequestParser
     {
+        internal static BatchRequestParser Instance = new BatchRequestParser(); // TODO arek - remove me
+
         // TODO sharding: CommandData seems to grew beyond the good practice of struct
         public struct CommandData
         {
@@ -75,7 +69,7 @@ namespace Raven.Server.Documents.Handlers
             public string DestinationName;
             public string ContentType;
             public AttachmentType AttachmentType;
-            public BatchHandler.MergedBatchCommand.AttachmentStream AttachmentStream; // used for bulk insert only
+            public MergedBatchCommand.AttachmentStream AttachmentStream; // used for bulk insert only
 
             #endregion Attachment
 
@@ -100,7 +94,6 @@ namespace Raven.Server.Documents.Handlers
             #endregion ravendata
         }
 
-        private static readonly CommandData[] Empty = new CommandData[0];
         private static readonly int MaxSizeOfCommandsInBatchToCache = 128;
 
         [ThreadStatic]
@@ -133,261 +126,7 @@ namespace Raven.Server.Documents.Handlers
             _cache.Push(cmds);
         }
 
-
-        public class DatabaseBatchCommandBuilder : AbstractBatchCommandBuilder<BatchHandler.MergedBatchCommand, DocumentsOperationContext>
-        {
-            private readonly DocumentDatabase _database;
-            public List<BatchHandler.MergedBatchCommand.AttachmentStream> AttachmentStreams;
-            public StreamsTempFile AttachmentStreamsTempFile;
-
-            public DatabaseBatchCommandBuilder(RequestHandler handler, DocumentDatabase database) : base(handler, database.Name, database.IdentityPartsSeparator)
-            {
-                _database = database;
-            }
-
-            public override async Task SaveStream(JsonOperationContext context, Stream input)
-            {
-                if (AttachmentStreams == null)
-                {
-                    AttachmentStreams = new List<BatchHandler.MergedBatchCommand.AttachmentStream>();
-                    AttachmentStreamsTempFile = _database.DocumentsStorage.AttachmentsStorage.GetTempFile("batch");
-                }
-
-                var attachmentStream = new BatchHandler.MergedBatchCommand.AttachmentStream
-                {
-                    Stream = AttachmentStreamsTempFile.StartNewStream()
-                };
-                attachmentStream.Hash = await AttachmentsStorageHelper.CopyStreamToFileAndCalculateHash(context, input, attachmentStream.Stream, _database.DatabaseShutdown);
-                await attachmentStream.Stream.FlushAsync();
-                AttachmentStreams.Add(attachmentStream);
-            }
-
-            public override async ValueTask<BatchHandler.MergedBatchCommand> GetCommandAsync(DocumentsOperationContext context)
-            {
-                await ExecuteGetIdentitiesAsync();
-
-                return new BatchHandler.MergedBatchCommand(_database)
-                {
-                    ParsedCommands = Commands,
-                    AttachmentStreams = AttachmentStreams,
-                    AttachmentStreamsTempFile = AttachmentStreamsTempFile,
-                    IsClusterTransaction = IsClusterTransactionRequest
-                };
-            }
-        }
-
-        public abstract class AbstractBatchCommandBuilder<TBatchCommand, TOperationContext>
-            where TBatchCommand : BatchHandler.IBatchCommand
-            where TOperationContext : JsonOperationContext
-        {
-            private readonly char _identityPartsSeparator;
-            protected readonly RequestHandler Handler;
-            private readonly string _database;
-            protected ServerStore ServerStore => Handler.ServerStore;
-
-            private int _index = -1;
-            private CommandData[] _commands = Empty;
-            public ArraySegment<CommandData> Commands => new ArraySegment<CommandData>(_commands, 0, _index + 1);
-
-            protected List<string> _identities;
-            protected List<int> _identityPositions;
-
-            public bool HasIdentities => _identities != null;
-            public bool IsClusterTransactionRequest;
-
-            protected AbstractBatchCommandBuilder(RequestHandler handler, string database, char identityPartsSeparator)
-            {
-                Handler = handler;
-                _database = database;
-                _identityPartsSeparator = identityPartsSeparator;
-            }
-
-            private void AddIdentity(JsonOperationContext ctx, ref CommandData command, int index)
-            {
-                _identities ??= new List<string>();
-                _identityPositions ??= new List<int>();
-
-                _identities.Add(command.Id);
-                _identityPositions.Add(index);
-
-                command.ChangeVector ??= ctx.GetLazyString("");
-            }
-
-            protected async ValueTask ExecuteGetIdentitiesAsync()
-            {
-                if (HasIdentities == false)
-                    return;
-
-                var newIds = await ServerStore.GenerateClusterIdentitiesBatchAsync(_database, _identities, RaftIdGenerator.NewId());
-                Debug.Assert(newIds.Count == _identities.Count);
-
-                for (var index = 0; index < _identityPositions.Count; index++)
-                {
-                    var value = _identityPositions[index];
-                    var cmd = _commands[value];
-
-                    cmd.Id = cmd.Id.Substring(0, cmd.Id.Length - 1) + _identityPartsSeparator + newIds[index];
-
-                    if (string.IsNullOrEmpty(cmd.ChangeVector) == false)
-                        ThrowInvalidUsageOfChangeVectorWithIdentities(cmd);
-
-                    // command it a struct so we need to copy it again
-                    _commands[value] = cmd;
-                }
-            }
-
-            public abstract Task SaveStream(JsonOperationContext context, Stream input);
-
-            public virtual Task<CommandData> ReadCommand(
-                JsonOperationContext ctx,
-                Stream stream,
-                JsonParserState state,
-                UnmanagedJsonParser parser,
-                JsonOperationContext.MemoryBuffer buffer,
-                BlittableMetadataModifier modifier,
-                CancellationToken token)
-            {
-                return ReadSingleCommand(ctx, stream, state, parser, buffer, modifier, token);
-            }
-
-            public async Task BuildCommandsAsync(JsonOperationContext context, Stream stream, char separator)
-            {
-                var state = new JsonParserState();
-                using (context.GetMemoryBuffer(out JsonOperationContext.MemoryBuffer buffer))
-                using (var parser = new UnmanagedJsonParser(context, state, "bulk_docs"))
-                /* In case we have a conflict between attachment with the same name we need attachment information from metadata */
-                /* we can't know from advanced if we will need this information so we save this for all batch commands */
-                using (var modifier = new BlittableMetadataModifier(context, legacyImport: false, readLegacyEtag: false, DatabaseItemType.Attachments))
-                {
-                    while (parser.Read() == false)
-                        await RefillParserBuffer(stream, buffer, parser);
-
-                    if (state.CurrentTokenType != JsonParserToken.StartObject)
-                        ThrowUnexpectedToken(JsonParserToken.StartObject, state);
-
-                    while (parser.Read() == false)
-                        await RefillParserBuffer(stream, buffer, parser);
-
-                    if (state.CurrentTokenType != JsonParserToken.String)
-                        ThrowUnexpectedToken(JsonParserToken.String, state);
-
-                    if (GetLongFromStringBuffer(state) != 8314892176759549763) // Commands
-                        ThrowUnexpectedToken(JsonParserToken.String, state);
-
-                    while (parser.Read() == false)
-                        await RefillParserBuffer(stream, buffer, parser);
-
-                    if (state.CurrentTokenType != JsonParserToken.StartArray)
-                        ThrowUnexpectedToken(JsonParserToken.StartArray, state);
-
-                    while (true)
-                    {
-                        while (parser.Read() == false)
-                            await RefillParserBuffer(stream, buffer, parser);
-
-                        if (state.CurrentTokenType == JsonParserToken.EndArray)
-                            break;
-
-                        _index++;
-                        if (_index >= _commands.Length)
-                        {
-                            _commands = IncreaseSizeOfCommandsBuffer(_index, _commands);
-                        }
-
-                        var commandData = await ReadCommand(context, stream, state, parser, buffer, modifier, Handler.AbortRequestToken);
-
-                        if (commandData.Type == CommandType.PATCH)
-                        {
-                            //TODO sharding: make it nicer
-                            commandData.PatchCommand =
-                                new PatchDocumentCommand(
-                                    context,
-                                    commandData.Id,
-                                    commandData.ChangeVector,
-                                    skipPatchIfChangeVectorMismatch: false,
-                                    (commandData.Patch, commandData.PatchArgs),
-                                    (commandData.PatchIfMissing, commandData.PatchIfMissingArgs),
-                                    commandData.CreateIfMissing,
-                                    isTest: false,
-                                    debugMode: false,
-                                    collectResultsNeeded: true,
-                                    returnDocument: commandData.ReturnDocument,
-                                    identityPartsSeparator: separator
-                                );
-                        }
-
-                        if (commandData.Type == CommandType.JsonPatch)
-                        {
-                            commandData.JsonPatchCommand = new JsonPatchCommand(
-                                commandData.Id,
-                                commandData.JsonPatchCommands,
-                                commandData.ReturnDocument,
-                                context);
-                        }
-
-                        if (commandData.Type == CommandType.BatchPATCH)
-                        {
-                            //TODO sharding: make it nicer
-                            commandData.PatchCommand =
-                                new BatchPatchDocumentCommand(
-                                    context,
-                                    commandData.Ids,
-                                    skipPatchIfChangeVectorMismatch: false,
-                                    (commandData.Patch, commandData.PatchArgs),
-                                    (commandData.PatchIfMissing, commandData.PatchIfMissingArgs),
-                                    commandData.CreateIfMissing,
-                                    isTest: false,
-                                    debugMode: false,
-                                    collectResultsNeeded: true
-                                );
-                        }
-
-                        if (IsIdentityCommand(ref commandData))
-                        {
-                            // queue identities requests in order to send them at once to the leader (using List for simplicity)
-                            AddIdentity(context, ref commandData, _index);
-                        }
-
-                        _commands[_index] = commandData;
-                    }
-
-                    if (await IsClusterTransaction(stream, parser, buffer, state))
-                        IsClusterTransactionRequest = true;
-                }
-            }
-
-            public static bool IsIdentityCommand(ref CommandData commandData)
-            {
-                return commandData.Type == CommandType.PUT && string.IsNullOrEmpty(commandData.Id) == false && commandData.Id[^1] == '|';
-            }
-
-            public async Task ParseMultipart(JsonOperationContext context, Stream stream, string contentType, char separator)
-            {
-                var boundary = MultipartRequestHelper.GetBoundary(
-                    MediaTypeHeaderValue.Parse(contentType),
-                    MultipartRequestHelper.MultipartBoundaryLengthLimit);
-                var reader = new MultipartReader(boundary, stream);
-                for (var i = 0; i < int.MaxValue; i++)
-                {
-                    var section = await reader.ReadNextSectionAsync().ConfigureAwait(false);
-                    if (section == null)
-                        break;
-
-                    var bodyStream = Handler.GetBodyStream(section);
-                    if (i == 0)
-                    {
-                        await BuildCommandsAsync(context, bodyStream, separator);
-                        continue;
-                    }
-
-                    await SaveStream(context, bodyStream);
-                }
-            }
-
-            public abstract ValueTask<TBatchCommand> GetCommandAsync(TOperationContext context);
-        }
-
-        private static async Task<bool> IsClusterTransaction(Stream stream, UnmanagedJsonParser parser, JsonOperationContext.MemoryBuffer buffer, JsonParserState state)
+        public async Task<bool> IsClusterTransaction(Stream stream, UnmanagedJsonParser parser, JsonOperationContext.MemoryBuffer buffer, JsonParserState state)
         {
             while (parser.Read() == false)
                 await RefillParserBuffer(stream, buffer, parser);
@@ -414,93 +153,17 @@ namespace Raven.Server.Documents.Handlers
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static unsafe long GetLongFromStringBuffer(JsonParserState state)
+        public static unsafe long GetLongFromStringBuffer(JsonParserState state)
         {
             return *(long*)state.StringBuffer;
         }
 
-        public class ReadMany : IDisposable
-        {
-            private readonly Stream _stream;
-            private readonly UnmanagedJsonParser _parser;
-            private readonly JsonOperationContext.MemoryBuffer _buffer;
-            private readonly JsonParserState _state;
-            private readonly CancellationToken _token;
-
-            public ReadMany(JsonOperationContext ctx, Stream stream, JsonOperationContext.MemoryBuffer buffer, CancellationToken token)
-            {
-                _stream = stream;
-                _buffer = buffer;
-                _token = token;
-
-                _state = new JsonParserState();
-                _parser = new UnmanagedJsonParser(ctx, _state, "bulk_docs");
-            }
-
-            public async Task Init()
-            {
-                while (_parser.Read() == false)
-                    await RefillParserBuffer(_stream, _buffer, _parser, _token);
-                if (_state.CurrentTokenType != JsonParserToken.StartArray)
-                {
-                    ThrowUnexpectedToken(JsonParserToken.StartArray, _state);
-                }
-            }
-
-            public void Dispose()
-            {
-                _parser.Dispose();
-            }
-
-            public Task<CommandData> MoveNext(JsonOperationContext ctx, BlittableMetadataModifier modifier)
-            {
-                if (_parser.Read())
-                {
-                    if (_state.CurrentTokenType == JsonParserToken.EndArray)
-                        return null;
-                    return ReadSingleCommand(ctx, _stream, _state, _parser, _buffer, modifier, _token);
-                }
-
-                return MoveNextUnlikely(ctx, modifier);
-            }
-
-            private async Task<CommandData> MoveNextUnlikely(JsonOperationContext ctx, BlittableMetadataModifier modifier)
-            {
-                do
-                {
-                    await RefillParserBuffer(_stream, _buffer, _parser, _token);
-                } while (_parser.Read() == false);
-                if (_state.CurrentTokenType == JsonParserToken.EndArray)
-                    return new CommandData { Type = CommandType.None };
-
-                return await ReadSingleCommand(ctx, _stream, _state, _parser, _buffer, modifier, _token);
-            }
-
-            public Stream GetBlob(long blobSize)
-            {
-                var bufferSize = _parser.BufferSize - _parser.BufferOffset;
-                var copy = ArrayPool<byte>.Shared.Rent(bufferSize);
-                var copySpan = new Span<byte>(copy);
-
-                _buffer.Memory.Memory.Span.Slice(_parser.BufferOffset, bufferSize).CopyTo(copySpan);
-
-                _parser.Skip(blobSize < bufferSize ? (int)blobSize : bufferSize);
-
-                return new LimitedStream(new ConcatStream(new ConcatStream.RentedBuffer
-                {
-                    Buffer = copy,
-                    Count = bufferSize,
-                    Offset = 0
-                }, _stream), blobSize, 0, 0);
-            }
-        }
-
-        public static async Task<CommandData> ReadAndCopySingleCommand(JsonOperationContext ctx,
+        public async Task<CommandData> ReadAndCopySingleCommand(JsonOperationContext ctx,
             Stream stream,
             JsonParserState state,
             UnmanagedJsonParser parser,
             JsonOperationContext.MemoryBuffer buffer,
-            ShardedBatchCommandBuilder.BufferedCommand bufferedCommand,
+            BufferedCommand bufferedCommand,
             BlittableMetadataModifier modifier,
             CancellationToken token)
         {
@@ -666,7 +329,7 @@ namespace Raven.Server.Documents.Handlers
             return commandData;
         }
 
-        private static async Task<CommandData> ReadSingleCommand(
+        public async Task<CommandData> ReadSingleCommand(
             JsonOperationContext ctx,
             Stream stream,
             JsonParserState state,
@@ -681,13 +344,18 @@ namespace Raven.Server.Documents.Handlers
                 ThrowUnexpectedToken(JsonParserToken.StartObject, state);
             }
 
+            CommandParsingObserver?.OnCommandStart(parser);
+
             while (true)
             {
                 while (parser.Read() == false)
                     await RefillParserBuffer(stream, buffer, parser, token);
 
                 if (state.CurrentTokenType == JsonParserToken.EndObject)
+                {
+                    CommandParsingObserver?.OnCommandEnd(parser);
                     break;
+                }
 
                 if (state.CurrentTokenType != JsonParserToken.String)
                 {
@@ -1058,7 +726,7 @@ namespace Raven.Server.Documents.Handlers
             return commandData;
         }
 
-        private static CommandData[] IncreaseSizeOfCommandsBuffer(int index, CommandData[] cmds)
+        public static CommandData[] IncreaseSizeOfCommandsBuffer(int index, CommandData[] cmds)
         {
             if (cmds.Length > MaxSizeOfCommandsInBatchToCache)
             {
@@ -1105,7 +773,7 @@ namespace Raven.Server.Documents.Handlers
             throw new InvalidOperationException($"Attachment PUT command must have a '{nameof(CommandData.Name)}' property");
         }
 
-        private static async Task<BlittableJsonReaderArray> ReadJsonArray(
+        private async Task<BlittableJsonReaderArray> ReadJsonArray(
             JsonOperationContext ctx,
             Stream stream,
             UnmanagedJsonParser parser,
@@ -1135,7 +803,7 @@ namespace Raven.Server.Documents.Handlers
             return reader;
         }
 
-        private static async Task<BlittableJsonReaderObject> ReadJsonObject(JsonOperationContext ctx, Stream stream, string id, UnmanagedJsonParser parser,
+        private async Task<BlittableJsonReaderObject> ReadJsonObject(JsonOperationContext ctx, Stream stream, string id, UnmanagedJsonParser parser,
             JsonParserState state, JsonOperationContext.MemoryBuffer buffer, IBlittableDocumentModifier modifier, CancellationToken token)
         {
             if (state.CurrentTokenType == JsonParserToken.Null)
@@ -1526,12 +1194,6 @@ namespace Raven.Server.Documents.Handlers
             return CommandType.None;
         }
 
-        private static void ThrowInvalidUsageOfChangeVectorWithIdentities(CommandData commandData)
-        {
-            throw new InvalidOperationException($"You cannot use change vector ({commandData.ChangeVector}) " +
-                                                $"when using identity in the document ID ({commandData.Id}).");
-        }
-
         private static unsafe void ThrowInvalidProperty(JsonParserState state, JsonOperationContext ctx)
         {
             throw new InvalidOperationException("Invalid property name: " +
@@ -1544,13 +1206,18 @@ namespace Raven.Server.Documents.Handlers
                                                   + ctx.AllocateStringValue(null, state.StringBuffer, state.StringSize));
         }
 
-        private static void ThrowUnexpectedToken(JsonParserToken jsonParserToken, JsonParserState state)
+        public static void ThrowUnexpectedToken(JsonParserToken jsonParserToken, JsonParserState state)
         {
             throw new InvalidOperationException("Expected " + jsonParserToken + " , but got " + state.CurrentTokenType);
         }
 
-        private static async Task RefillParserBuffer(Stream stream, JsonOperationContext.MemoryBuffer buffer, UnmanagedJsonParser parser, CancellationToken token = default)
+
+        public AbstractBatchCommandParsingObserver CommandParsingObserver { get; set; }
+
+        public async Task RefillParserBuffer(Stream stream, JsonOperationContext.MemoryBuffer buffer, UnmanagedJsonParser parser, CancellationToken token = default)
         {
+            CommandParsingObserver?.OnParserBufferRefill(parser);
+
             // Although we using here WithCancellation and passing the token,
             // the stream will stay open even after the cancellation until the entire server will be disposed.
             var read = await stream.ReadAsync(buffer.Memory.Memory, token);
