+++ src/Raven.Server/Documents/TransactionCommands/JsonPatchCommand.cs
@@ -6,6 +6,7 @@ using Raven.Client;
 using Raven.Client.Documents.Commands.Batches;
 using Raven.Client.Documents.Operations;
 using Raven.Server.Documents.Handlers;
+using Raven.Server.Documents.Handlers.Batches;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
