+++ src/Raven.Server/ServerWide/Commands/ClusterTransactionCommand.cs
@@ -8,6 +8,8 @@ using Raven.Client.ServerWide;
 using Raven.Client.Util;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Handlers;
+using Raven.Server.Documents.Handlers.Batches;
+using Raven.Server.Documents.Sharding;
 using Raven.Server.Json;
 using Raven.Server.Rachis;
 using Raven.Server.ServerWide.Context;
