+++ src/Raven.Server/SqlMigration/SqlMigrationWriter.cs
@@ -5,7 +5,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Raven.Client.Util;
 using Raven.Server.Documents;
-using Raven.Server.Documents.Handlers;
+using Raven.Server.Documents.Handlers.Batches;
+using Raven.Server.Documents.Handlers.Batches.Commands;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using CommandType = Raven.Client.Documents.Commands.Batches.CommandType;
@@ -15,7 +16,7 @@ namespace Raven.Server.SqlMigration
     internal class SqlMigrationWriter : IDisposable
     {
         private readonly DocumentsOperationContext _context;
-        private BatchHandler.MergedBatchCommand _command;
+        private MergedBatchCommand _command;
         private readonly int _batchSize;
 
         public SqlMigrationWriter(DocumentsOperationContext context, int batchSize)
@@ -27,7 +28,7 @@ namespace Raven.Server.SqlMigration
         }
 
         private readonly List<BatchRequestParser.CommandData> _commands = new List<BatchRequestParser.CommandData>();
-        private readonly List<BatchHandler.MergedBatchCommand.AttachmentStream> _attachmentStreams = new List<BatchHandler.MergedBatchCommand.AttachmentStream>();
+        private readonly List<MergedBatchCommand.AttachmentStream> _attachmentStreams = new List<MergedBatchCommand.AttachmentStream>();
         private readonly List<IDisposable> _toDispose = new List<IDisposable>();
 
         public async Task InsertDocument(BlittableJsonReaderObject document, string id, Dictionary<string, byte[]> attachments)
@@ -47,7 +48,7 @@ namespace Raven.Server.SqlMigration
                 _toDispose.Add(memoryStream);
                 _toDispose.Add(stream);
 
-                var attachmentStream = new BatchHandler.MergedBatchCommand.AttachmentStream
+                var attachmentStream = new MergedBatchCommand.AttachmentStream
                 {
                     Stream = stream,
                     Hash = await AttachmentsStorageHelper.CopyStreamToFileAndCalculateHash(_context, memoryStream, stream, _context.DocumentDatabase.DatabaseShutdown) //TODO: do we need it?
@@ -76,7 +77,7 @@ namespace Raven.Server.SqlMigration
                 return;
 
             _command.ParsedCommands = new ArraySegment<BatchRequestParser.CommandData>(_commands.ToArray(), 0, _commands.Count);
-            _command.AttachmentStreams = new List<BatchHandler.MergedBatchCommand.AttachmentStream>(_attachmentStreams);
+            _command.AttachmentStreams = new List<MergedBatchCommand.AttachmentStream>(_attachmentStreams);
 
             try
             {
@@ -99,7 +100,7 @@ namespace Raven.Server.SqlMigration
         private void Reset()
         {
             _command?.Dispose();
-            _command = new BatchHandler.MergedBatchCommand(_context.DocumentDatabase)
+            _command = new MergedBatchCommand(_context.DocumentDatabase)
             {
                 AttachmentStreamsTempFile = _context.DocumentDatabase.DocumentsStorage.AttachmentsStorage.GetTempFile("put")
             };
