+++ src/Raven.Studio/typescript/commands/database/documents/executeBulkDocsCommand.ts
@@ -3,12 +3,12 @@ import database = require("models/resources/database");
 import endpoints = require("endpoints");
 
 class executeBulkDocsCommand extends commandBase {
-    constructor(public docs: Partial<Raven.Server.Documents.Handlers.BatchRequestParser.CommandData>[], private db: database, private transactionMode: Raven.Client.Documents.Session.TransactionMode = "SingleNode") {
+    constructor(public docs: Partial<Raven.Server.Documents.Handlers.Batches.BatchRequestParser.CommandData>[], private db: database, private transactionMode: Raven.Client.Documents.Session.TransactionMode = "SingleNode") {
         super();
     }
 
-    execute(): JQueryPromise<resultsDto<Raven.Server.Documents.Handlers.BatchRequestParser.CommandData>> {
-        return this.post<Raven.Server.Documents.Handlers.BatchRequestParser.CommandData[]>(endpoints.databases.batch.bulk_docs, JSON.stringify({ Commands: this.docs, TransactionMode: this.transactionMode }), this.db);
+    execute(): JQueryPromise<resultsDto<Raven.Server.Documents.Handlers.Batches.BatchRequestParser.CommandData>> {
+        return this.post<Raven.Server.Documents.Handlers.Batches.BatchRequestParser.CommandData[]>(endpoints.databases.batch.bulk_docs, JSON.stringify({ Commands: this.docs, TransactionMode: this.transactionMode }), this.db);
     }
 }
