+++ test/SlowTests/Client/BulkInserts.cs
@@ -1,4 +1,7 @@
-﻿using System.Threading.Tasks;
+﻿
+using System.IO.Compression;
+using System.Threading.Tasks;
+using FastTests;
 using Tests.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
@@ -11,12 +14,13 @@ namespace SlowTests.Client
         {
         }
 
-        [Fact]
-        public async Task Simple_Bulk_Insert_With_Ssl()
+        [Theory]
+        [RavenData(false, DatabaseMode = RavenDatabaseMode.All)]
+        public async Task Simple_Bulk_Insert_With_Ssl(RavenTestBase.Options options)
         {
             using (var x = new FastTests.Client.BulkInserts(Output))
             {
-                await x.Simple_Bulk_Insert(useSsl: true);
+                await x.Simple_Bulk_Insert(options, useSsl: true, compressionLevel: CompressionLevel.NoCompression);
             }
         }
     }
